# docs/swagger.yaml
basePath: /
definitions:
  core.APIResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
      timestamp:
        type: string
    type: object
...
  core.ErrorResponse:
    properties:
      details: {}
      error:
        type: string
      message:
        type: string
      status:
        type: string
      timestamp:
        type: string
    type: object
...
  core.MidtransPaymentRequest:
    properties:
      amount:
        type: number
      customer_email:
        type: string
      customer_phone:
        type: string
      order_id:
        type: string
    required:
    - amount
    - customer_email
    - customer_phone
    - order_id
    type: object
...
  core.PaginatedResponse:
    properties:
      data: {}
      message:
        type: string
      pagination:
        $ref: '#/definitions/core.PaginationResponse'
      status:
        type: string
      timestamp:
        type: string
    type: object
  core.PaginationResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  core.PaymentMethod:
    enum:
    - cash
    - midtrans
    - gopay
    - qris
    - qris
    - transfer
    - bank_transfer
    - mandiri_echannel
    type: string
    x-enum-comments:
      PaymentMethodQris: Alias for QRIS
    x-enum-descriptions:
    - ""
    - ""
    - ""
    - ""
    - Alias for QRIS
    - ""
    - ""
    - ""
    x-enum-varnames:
    - PaymentMethodCash
    - PaymentMethodMidtrans
    - PaymentMethodGopay
    - PaymentMethodQRIS
    - PaymentMethodQris
    - PaymentMethodTransfer
    - PaymentMethodBankTransfer
    - PaymentMethodMandiriEchannel
  core.PaymentRequest:
    properties:
      amount:
        type: number
      notes:
        type: string
      order_id:
        type: string
      payment_method:
        $ref: '#/definitions/core.PaymentMethod'
    required:
    - amount
    - order_id
    - payment_method
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@iphoneservice.com
    name: iPhone Service API Support
    url: https://iphoneservice.com
  description: Backend API for iPhone Service Point of Sales system supporting 50
    branches across Indonesia
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: iPhone Service POS API
  version: "1.0"
paths:
...
  /payments:
    get:
      consumes:
      - application/json
      description: Get list of payments with pagination and filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by order ID
        in: query
        name: order_id
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by payment method
        in: query
        name: payment_method
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.PaginatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: List payments
      tags:
      - payments
    post:
      consumes:
      - application/json
      description: Create a new payment for an order
      parameters:
      - description: Payment data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/core.PaymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/core.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new payment
      tags:
      - payments
  /payments/{id}:
    get:
      consumes:
      - application/json
      description: Get payment details by ID
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: Get payment by ID
      tags:
      - payments
  /payments/{id}/status:
    put:
      consumes:
      - application/json
      description: Update the status of a payment
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      - description: Payment Status
        in: query
        name: status
        required: true
        type: string
      - description: Transaction ID
        in: query
        name: transaction_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update payment status
      tags:
      - payments
  /payments/invoice/{invoiceNumber}:
    get:
      consumes:
      - application/json
      description: Get payment details by invoice number
      parameters:
      - description: Invoice Number
        in: path
        name: invoiceNumber
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: Get payment by invoice number
      tags:
      - payments
  /payments/midtrans:
    post:
      consumes:
      - application/json
      description: Process payment through Midtrans gateway
      parameters:
      - description: Midtrans payment data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/core.MidtransPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Process Midtrans payment
      tags:
      - payments
  /payments/order/{orderId}:
    get:
      consumes:
      - application/json
      description: Get all payments for a specific order
      parameters:
      - description: Order ID
        in: path
        name: orderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: Get payments by order
      tags:
      - payments
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"